Ec report

See merge request codeaster/src!250

+++ bibfor/algorith/statim.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2024 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -102,8 +102,8 @@ subroutine statim(nbobst, nbpt, temps, fcho, vgli, &
     ibid = 0
 !
     dt = (temps(nbpt)-temps(1))/(nbpt-1)
-    idebut = int((tdebut-temps(1))/dt)+1
-    ifin = min(int((tfin-temps(1))/dt)+1, nbpt)
+    idebut = nint((tdebut-temps(1))/dt)+1
+    ifin = min(nint((tfin-temps(1))/dt)+1, nbpt)
     nbpas = ifin-idebut+1
 !
     if (nbloc .eq. 0) nbloc = 1
